+++ src/shared/queue/queue.module.ts
@@ -1,15 +1,16 @@
 import { Module } from '@nestjs/common';
 import { ApiModule } from '../api';
 import { DbModule } from '../db';
+import { GameModule } from '../game/game.module';
 import { BattleProcessor } from './processors';
-import { CardPollProcessor } from './processors/card.processor';
+import { CardProcessor } from './processors/card.processor';
 import { MigrateProcessor } from './processors/migrate.processor';
 import { QueueEventsService } from './queue-events.service';
 @Module({
-  imports: [ApiModule, DbModule],
+  imports: [ApiModule, DbModule, GameModule],
   providers: [
     BattleProcessor,
-    CardPollProcessor,
+    CardProcessor,
     MigrateProcessor,
     QueueEventsService,
   ],