+++ src/util/forms/index.ts
@@ -2,3 +2,4 @@ export * from './battle-form';
 export * from './collection-form';
 export * from './history-form';
 export * from './leaderboard-form';
+export * from './marketplace-form';